+++ lucabook/lib/luca_book.rb
@@ -9,6 +9,7 @@ module LucaBook
   autoload :Import, 'luca_book/import'
   autoload :Journal, 'luca_book/journal'
   autoload :List, 'luca_book/list'
+  autoload :ListByHeader, 'luca_book/list_by_header'
   autoload :Setup, 'luca_book/setup'
   autoload :State, 'luca_book/state'
   autoload :Util, 'luca_book/util'